base: Expand @core, drop some bits from it
For my libhif (librepo/hawkey) branch of rpm-ostree, we'd need to pull
in libcomps in order to expand @core.  Because I have some outstanding
issues with that code
<https://github.com/midnightercz/libcomps/issues/22>, let's just
expand @core manually.

A major bonus of this is it allows us to drop yum and some other
components like authconfig and firewalld.