treecompose: Use an external webserver with dynamic port
I was having trouble with imagefactory accessing the content; it
turned out to be firewalld on the host blocking access to the port.

In the process of debugging that I switched to using an external
webserver as then I could just strace -p it to see what was going
on

A nice benefit of ostree trivial-httpd is it dynamically allocates an
*unused* port.  That way we don't need to hardcode 8000.

This patch changes the kickstart file to use that dynamic port.

Furthermore, I changed the IP address range to match the observed
range I'm getting here.  However...that may be a local libvirt change.

I need to figure out whether there's a reliable way to communciate the
host IP address/location into the kickstart we sent to the VM.