manifest: Drop heat-cfntools
Right now, heat-cfntools has a rather large dependency chain as it's
attempting to wrap a small subset of the known package systems and
language installers in the world.

It pulls in yum and ruby, among other things.

The problem with yum is not size exactly, but the fact that its
presence confuses management tools into thinking it can operate to
install packages.  One can use yum in a read-only mode, but on the
other hand, one can just run it in a container too.

See: https://bugzilla.redhat.com/show_bug.cgi?id=1130964