treecompose-post: Fix locale dropping by forcing text grep
It looks like glibc's locales at some point gained "français" as a
locale with UTF-8 encoding.  When we're doing the processing here, a
locale is not set, so grep doesn't recognize the UTF-8 and just
outputs "(Binary file matches)" which obviously breaks things.

Fix this by forcing grep to treat input as text.