manifest: Add screen
This is a drive by commit with no review, I know...but I keep getting
complaints about this.  It's a tiny binary, and it doesn't make a lot
of sense to run tmux/screen inside a container, as normally you want
them to be bound to the lifecycle of your login, not dockerd.