add oci-register-machine and oci-systemd-hook
The ostree tree from the initial release of fedora atomic 24 (24.39)
included oci-register-machine and oci-systemd-hook, which centos and
rhel atomic hosts also include. These hooks deal with systemd
integration:
(http://rhelblog.redhat.com/2016/06/30/whats-new-in-red-hat-enterprise-linux-atomic-host-7-2-5/)

These packages were pulled in as a dependency of docker-1.10.3-9, but
docker-1.10.3-24 does not, for some reason, pull them in -- both
specify these as "Recommends," and that has not changed. In rhel and
centos, these oci hook pages are "Requires" for docker, so it's unclear
whether this issue would affect them.

This commit adds these oci hooks in the json, to restore these systemd
integration features to fedora atomic 24.