treecompose-post.sh: copy, don't move the locale archive
Most processes, including bash, will have locale-archive mmap'ed.
Renaming it to .tmpl will cause it to get truncated by
build-locale-archive. We should make sure that it remains untouched
so that bash doesn't die with SIGBUS. OTOH, build-locale-archive knows
to unlink and create a new inode so as to not affect already mmap'ed
versions.